+++ sdk/core/core-client-rest/src/helpers/isBinaryBody.ts
@@ -1,15 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import type { NodeJSReadableStream } from "@azure/core-rest-pipeline";
 import { isReadableStream } from "./isReadableStream.js";
 
 export function isBinaryBody(
   body: unknown,
 ): body is
   | Uint8Array
-  | NodeJS.ReadableStream
+  | NodeJSReadableStream
   | ReadableStream<Uint8Array>
-  | (() => NodeJS.ReadableStream)
+  | (() => NodeJSReadableStream)
   | (() => ReadableStream<Uint8Array>)
   | Blob {
   return (